requirements.txt
Flask==2.3.2
gunicorn==22.0.0
Pillow==12.1.1
numpy>=2.0.0